Fix webpack related issues
