+++ app/javascript/components/App.js
@@ -1,8 +1,7 @@
-
-
 import React from 'react'
 import { Provider } from 'react-redux';
 import {  BrowserRouter, Routes, Route } from 'react-router-dom';
+// import { Provider } from 'react-redux';
 import Greeting from './Greeting';
 import Header from './Header';
 import Home from './Home';
@@ -10,15 +9,15 @@ import store from '../redux/ConfigureStore';
 
 const App = () => {
   return (
-    <Provider store={store}>
+      // <Provider store={store}> 
       <BrowserRouter>
       <Header />
         <Routes>
-          <Route path="/" element={(<Home />)} />
-          <Route path="/greeting" element={(<Greeting />)} />
+          <Route exact path="/" element={ (<Home />) } />
+          <Route exact path="/greeting" element={ (<Greeting />) } />
         </Routes>
       </BrowserRouter>
-    </Provider>
+    // </Provider>
   )
 }
