+++ app/javascript/components/Greeting.js
@@ -13,9 +13,9 @@ const Greeting =() => {
     window.location.reload(false);
   }
     return (
-        <div className="">
+        <div className="mt-20 flex flex-col items-center">
           <p className=" font-bold">{ greeting }</p>
-          <button onClick={refreshPage} className="">Get another greeting</button>
+          <button onClick={refreshPage} className="bg-sky-500 p-2 mt-4 rounded text-slate-50">Get another greeting</button>
       </div>
     );
 }
