+++ babel.config.js
@@ -22,10 +22,8 @@ module.exports = function(api) {
         {
           targets: {
             node: 'current'
-          },
-          modules: 'commonjs'
-        },
-        '@babel/preset-react'
+          }
+        }
       ],
       (isProductionEnv || isDevelopmentEnv) && [
         '@babel/preset-env',
@@ -35,13 +33,6 @@ module.exports = function(api) {
           corejs: 3,
           modules: false,
           exclude: ['transform-typeof-symbol']
-        }
-      ],
-      [
-        '@babel/preset-react',
-        {
-          development: isDevelopmentEnv || isTestEnv,
-          useBuiltIns: true
         }
       ]
     ].filter(Boolean),
@@ -77,21 +68,13 @@ module.exports = function(api) {
       [
         '@babel/plugin-transform-runtime',
         {
-          helpers: false,
-          regenerator: true,
-          corejs: false
+          helpers: false
         }
       ],
       [
         '@babel/plugin-transform-regenerator',
         {
           async: false
-        }
-      ],
-      isProductionEnv && [
-        'babel-plugin-transform-react-remove-prop-types',
-        {
-          removeImport: true
         }
       ]
     ].filter(Boolean)
