+++ package.json
@@ -9,19 +9,19 @@
   "dependencies": {
     "@babel/preset-react": "^7.18.6",
     "@rails/webpacker": "5.4.3",
+    "axios": "^0.25.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "babel-polyfill": "^6.26.0",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
-    "axios": "^0.25.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.1",
     "react-router-dom": "^6.3.0",
+    "react_ujs": "^2.6.2",
     "redux": "^4.2.0",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.4.1",
     "reselect": "^4.1.5",
-    "react_ujs": "^2.6.2",
     "webpack": "^4.46.0",
     "webpack-cli": "^3.3.12"
   }
